not sure

+++ src/main.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include "utils.h"
 
 int main(int argc, char *argv[]) {
-    int globalPassExists = checkGlobalPass();
+    uint8_t globalPassExists = checkGlobalPass();
     if (argc != 2) {
         printf(
             "Arguments do no match\n"
@@ -17,7 +18,7 @@ int main(int argc, char *argv[]) {
     if (!globalPassExists) {
         setGlobalPass();
     }
-    
+
     if (strcmp(argv[1], HELP) == 0) {
         helpScreen();
     } else if (strcmp(argv[1], GEN) == 0) {
